refac: move file page index
